% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clparlysearch.R
\docType{package}
\name{clparlysearch}
\alias{clparlysearch}
\title{clparlysearch: Download data on Members' activities in Parliament}
\description{
Package to download data on Members' Parliamentary activities
}
