% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_csv}
\alias{download_csv}
\title{\code{download_csv} Downloads a csv from a url}
\usage{
download_csv(url)
}
\value{
A tibble
}
\description{
\code{download_csv} Downloads a csv from a url
}
\keyword{internal}
