% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_proceedings.R
\name{fetch_all_proceedings}
\alias{fetch_all_proceedings}
\title{\code{fetch_all_proceedings} Download data for proceedings}
\usage{
fetch_all_proceedings(session_start, session_end)
}
\arguments{
\item{session_start}{The start year of the Session (yy)}

\item{session_end}{The end year of the Session (yy)}
}
\value{
A dataframe including contributions data for proceedings
}
\description{
\code{fetch_all_proceedings} Download data for proceedings
}
