% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_chamber_contributions.R
\name{fetch_chamber_contributions}
\alias{fetch_chamber_contributions}
\title{Get contributions data for a defined period}
\usage{
fetch_chamber_contributions(start_date, end_date)
}
\arguments{
\item{start_date}{The start date of the period for which contributions are downloaded ("yyyy-mm-dd")}

\item{end_date}{The end date of the period for which contributions are downloaded ("yyyy-mm-dd")}
}
\value{
A tibble containing contributions along with details such as member, date, and subject
}
\description{
\code{fetch_chamber_contributions} Downloads Members' contributions spoken in the Commons Chamber for a defined period, excluding Question not (yet) asked
}
