% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_debate.R
\name{fetch_debate}
\alias{fetch_debate}
\title{Get contributions data for a specific debate}
\usage{
fetch_debate(title, date)
}
\arguments{
\item{title}{The title of the debate (e.g. "Brexit transition")}

\item{date}{The date of the debate ("yyyy-mm-dd")}
}
\value{
A tibble containing contributions along with details such as Member, date, and subject
}
\description{
\code{fetch_debate} Downloads Members' contributions to a specific debate
}
