% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_petitions.R
\name{fetch_petitions}
\alias{fetch_petitions}
\title{Get data on petitions by Member, by Session}
\usage{
fetch_petitions(session_start, session_end)
}
\arguments{
\item{session_start}{The start year of the session (yy)}

\item{session_end}{The end year of the session (yy)}
}
\value{
A dataframe including data on petitions
}
\description{
\code{fetch_petitions} Downloads data on petitions introduced by Member, by Session
}
