% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_pmbs.R
\name{fetch_pmbs}
\alias{fetch_pmbs}
\title{Get data on PMBs by Session}
\usage{
fetch_pmbs(session_start, session_end)
}
\arguments{
\item{session_start}{The start year of the Session (yy)}

\item{session_end}{The end year of the Session (yy)}
}
\value{
A dataframe containing data on bills introduced
}
\description{
\code{fetch_pmbs} Get data on Private Members' Bills introduced by Member by Session
}
