% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_dates}
\alias{generate_dates}
\title{\code{generate_dates} Generates a list of dates to map over}
\usage{
generate_dates(start_date, end_date)
}
\value{
A list of weeks in a specified period broken down into day, month, year
}
\description{
\code{generate_dates} Generates a list of dates to map over
}
\keyword{internal}
