% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_members}
\alias{get_members}
\title{Helper functions}
\usage{
get_members()
}
\value{
A tibble data on all current Members including MNIS id, name, gender, constituency, start date and party
}
\description{
\code{get_members} Downloads data on all current Members
}
\keyword{internal}
