% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_proceedings.R
\name{get_urls}
\alias{get_urls}
\title{Get data on proceedings initiated by Members by Session}
\usage{
get_urls(session_start, session_end)
}
\arguments{
\item{session_start}{The start year of the Session (yy)}

\item{session_end}{The end year of the Session (yy)}
}
\value{
A list of urls
}
\description{
\code{get_urls} Generate urls to download data
}
\keyword{internal}
