% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_contributions.R
\name{process_contributions}
\alias{process_contributions}
\title{Process contributions data}
\usage{
process_contributions(contributions)
}
\arguments{
\item{contributions}{A dataframe containing Members' contribution data}
}
\value{
A dataframe containing contributions attributed to single members, including the number of words per contribution
}
\description{
\code{process_contributions} Processes Members' contributions data
}
